Move user input to separate module
