+++ user_input_module/__init__.py
@@ -1,0 +1,1 @@
+from .user_input import InputHandler
